refactor(api): move UserWithActivity model
