+++ api/src/identity-access-management/domain/read-models/UserWithActivity.js
