+++ api/tests/unit/infrastructure/serializers/jsonapi/user-with-activity-serializer_test.js
@@ -1,6 +1,6 @@
-import { UserWithActivity } from '../../../../../lib/domain/read-models/UserWithActivity.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/user-with-activity-serializer.js';
 import { User } from '../../../../../src/identity-access-management/domain/models/User.js';
+import { UserWithActivity } from '../../../../../src/identity-access-management/domain/read-models/UserWithActivity.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | user-with-activity-serializer', function () {
